feat(server,data): tighten types and parallelise derived-table setup
1. ranking_factors: rank_* INTEGER → SMALLINT
   `data/processors/search.py` caps the values at 1100 / 900 / 99 with a
   combined formula that fits comfortably in i16. Saves 10 bytes per row
   and makes the bound explicit at the schema layer.

2. main.rs: run the 9 derived-table loaders in parallel via JoinSet.
   They have no inter-dependencies (each FKs back to `de` or `en`, never
   to another derived table). Pool max bumped to 20 to fit 9 concurrent
   txs plus headroom.

3. test_get_handler_unchanged: fix the request URI — get_handler is
   mounted at /api/locations/{id}, the test was hitting /{key} and
   getting 404 on every row.

+++ data/external/schemas/ranking_factors.py
@@ -4,11 +4,13 @@ import polars as pl
 
 class RankingFactorsSchema(dy.Schema):
     id = dy.String(primary_key=True, nullable=False)
-    rank_type = dy.Int32(nullable=True)
-    rank_combined = dy.Int32(nullable=True)
-    rank_usage = dy.Int32(nullable=True)
-    rank_custom = dy.Int32(nullable=True)
-    rank_boost = dy.Int32(nullable=True)
+    # SMALLINT bounds enforced by `data/processors/search.py` (rank_type<=1100,
+    # rank_usage<=900, rank_boost<=99); combined fits comfortably in i16.
+    rank_type = dy.Int16(nullable=True)
+    rank_combined = dy.Int16(nullable=True)
+    rank_usage = dy.Int16(nullable=True)
+    rank_custom = dy.Int16(nullable=True)
+    rank_boost = dy.Int16(nullable=True)
 
     @dy.rule()
     def id_non_empty(cls) -> pl.Expr:
